release_to: Fuze test and publish scripts
