+++ config/release-management-tools.json
@@ -7,7 +7,6 @@
     "list_accepted_pkgs",
     "jira_epics",
     "sle_build",
-    "release_to_test",
-    "release_to_publish"
+    "release_to"
   ]
 }
